+++ bandlog-team/src/main/java/net/effize/bandlog/team/port/UserPort.java
@@ -1,10 +1,10 @@
-package net.effize.bandlog.team.repository;
+package net.effize.bandlog.team.port;
 
 import net.effize.bandlog.team.model.User;
 import net.effize.bandlog.team.model.UserId;
 
 import java.util.List;
 
-public interface UserRepository {
+public interface UserPort {
     List<User> findAllByIdIn(List<UserId> ids);
 }
