+++ bandlog-team/src/main/java/net/effize/bandlog/team/service/TeamQueryService.java
@@ -7,7 +7,7 @@ import net.effize.bandlog.team.model.Team;
 import net.effize.bandlog.team.model.TeamId;
 import net.effize.bandlog.team.model.User;
 import net.effize.bandlog.team.model.UserId;
-import net.effize.bandlog.team.repository.UserRepository;
+import net.effize.bandlog.team.port.UserPort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,11 +17,11 @@ import java.util.stream.Collectors;
 @Service
 public class TeamQueryService {
     private final TeamService teamService;
-    private final UserRepository userRepository;
+    private final UserPort userPort;
 
-    public TeamQueryService(TeamService teamService, UserRepository userRepository) {
+    public TeamQueryService(TeamService teamService, UserPort userPort) {
         this.teamService = teamService;
-        this.userRepository = userRepository;
+        this.userPort = userPort;
     }
 
     public TeamInfoResponse teamInfo(UserId authUserId, Long teamId) {
@@ -30,7 +30,7 @@ public class TeamQueryService {
         long meCount = members.stream().filter(member -> member.userId().equals(authUserId)).count();
         if (meCount <= 0) throw new IllegalStateException("User is not a member of the team");
 
-        List<User> users = userRepository.findAllByIdIn(
+        List<User> users = userPort.findAllByIdIn(
                 members
                         .stream()
                         .map(Member::userId)
@@ -63,7 +63,7 @@ public class TeamQueryService {
                 teams.stream().map(team -> {
                     List<Member> members = teamService.membersOf(team);
 
-                    List<User> users = userRepository.findAllByIdIn(
+                    List<User> users = userPort.findAllByIdIn(
                             members
                                     .stream()
                                     .map(Member::userId)
